+++ Framework/Scene.cs
@@ -386,7 +386,9 @@
         /// <param name="component">The component.</param>
         /// <returns>A value indicating whether or not the child was removed.</returns>
         public bool RemoveChild(BaseComponent component) {
-            return this._components.Remove(component) || this.ComponentsForSaving.Remove(component);
+            var result = this._components.Remove(component);
+            result = this.ComponentsForSaving.Remove(component) || result;
+            return result;
         }
 
         /// <summary>